料兵家常验，则太史多行占，我算世间到底不容圆。
